Move GT7 agent to gt7 folder. Clean-up and modify pipelines for Nuitka build
